Added CRUD with REST

+++ src/main/java/com/simbirsoft/itplace/repository/PersonalDataRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Denis on 08.04.2019.
  */
-public interface PersonalDataRepository extends JpaRepository<PersonalData, Integer> {
+public interface PersonalDataRepository extends JpaRepository<PersonalData, Long> {
 }
